Исправил в dtoAds поле price, по openApi там должно быть Long, думал из-за этого проблемы в json, но это не помогло. Думаю возможно все дело в спринг секюрити.. В вс буду изучать смотреть ролики))
Ну и добавил сервис с картинками, но там пока тоже что-то не пошло.

!!!Если запускать эту ветку не забудь удалить папку target (в ней сохраняются старые бины и из-за этого могут быть конфликты)

+++ src/main/java/ru/skypro/homework/dto/AdsDto.java
@@ -10,6 +10,6 @@ public class AdsDto {
     private String title;
     private String description;
     private String image;
-    private Double price;
+    private Long price;
 
 }
